鍵盤・指板のクリックで音を鳴らす、テンポのDAW同期（DAW上では既定でオン）
- 鍵盤（C3〜C6）・指板（E2〜）を押すとその音を試聴（playNotes）
- processBlock で DAW のテンポを取得し、hostTempo イベントで UI へ通知
- 「DAW同期」ボタン（Standaloneでは非表示）。同期中は試聴・ドラッグ・保存に DAW のテンポを使用
- MIDI 出力が小数テンポ（20〜300）を受けるように

+++ Source/MidiExport.cpp
@@ -45,7 +45,7 @@ std::vector<int> voicing (const Chord& chord)
     return notes;
 }
 
-juce::MemoryBlock buildMidi (const std::vector<Chord>& chords, int bpm)
+juce::MemoryBlock buildMidi (const std::vector<Chord>& chords, double bpm)
 {
     bpm = juce::jlimit (minBpm, maxBpm, bpm);
     const auto usPerQuarter = (juce::uint32) juce::roundToInt (60000000.0 / bpm);
@@ -119,7 +119,7 @@ juce::String safeFileName (const juce::String& name)
     return out.isEmpty() ? juce::String ("chord") : out;
 }
 
-juce::File writeTempFile (const std::vector<Chord>& chords, int bpm, const juce::String& name)
+juce::File writeTempFile (const std::vector<Chord>& chords, double bpm, const juce::String& name)
 {
     auto dir = juce::File::getSpecialLocation (juce::File::tempDirectory).getChildFile ("GodokenNavigator");
